Refactor ipc code in main to a different file
