+++ src/main/main.ts
@@ -9,15 +9,14 @@
  * `./src/main.js` using webpack. This gives us some performance wins.
  */
 import path from 'path';
-import { app, BrowserWindow, shell, ipcMain } from 'electron';
+import { app, BrowserWindow, shell } from 'electron';
 import { autoUpdater } from 'electron-updater';
 import log from 'electron-log';
 import AppDataSource from '../data-source';
 import MenuBuilder from './menu';
 import { resolveHtmlPath } from './util';
-import Procedures from '../db/Procedures';
 import 'reflect-metadata';
-import ConnectionService from '../db/service/ConnectionService';
+import './ipc';
 
 class AppUpdater {
   constructor() {
@@ -28,32 +27,6 @@ class AppUpdater {
 }
 
 let mainWindow: BrowserWindow | null = null;
-
-ipcMain.on('ipc-example', async (event, arg) => {
-  const msgTemplate = (pingPong: string) => `IPC test: ${pingPong}`;
-  console.log(msgTemplate(arg));
-  event.reply('ipc-example', msgTemplate('pong'));
-});
-
-ipcMain.handle('procedures:listProcedures', () => {
-  return new Procedures().getProceduresForDB(['React']);
-});
-
-ipcMain.handle('procedures:listDatabases', () => {
-  return new Procedures().getDatabases();
-});
-
-ipcMain.handle('procedures:getProcedure', async (_event, ...args) => {
-  return new Procedures().fetchContent('React', args[0]);
-});
-
-ipcMain.handle('connections:create', (_event, ...args) => {
-  return new ConnectionService().create(args[0]);
-});
-
-ipcMain.handle('connections:select', (_event, ...args) => {
-  return new ConnectionService().select(args[0]);
-});
 
 if (process.env.NODE_ENV === 'production') {
   const sourceMapSupport = require('source-map-support');
